Widen get_total_latency() return type to double
After the RED-191460 fix, totals::m_latency is a double. The
get_total_latency() getter still returned unsigned long int and
silently narrowed any fractional value to 0 or 1. The getter is
currently only invoked during the live-progress phase where
m_latency holds an integer μs accumulator, so no current caller
truncates — but the type mismatch is the same bug class that
RED-191460 introduced and is one summarize() call away from
re-triggering it.

Propagate double through:
  run_stats::get_total_latency           returns double
  client_group::get_total_latency        returns double
  memtier_benchmark.cpp local accumulator  becomes double

The display arithmetic (avg_latency, cur_latency) was already
done in double, so no precision is lost downstream.

+++ run_stats.h
@@ -257,7 +257,7 @@ public:
     unsigned long int get_duration_usec(void);
     unsigned long int get_total_bytes(void);
     unsigned long int get_total_ops(void);
-    unsigned long int get_total_latency(void);
+    double get_total_latency(void);
     unsigned long int get_total_connection_errors(void);
 
     // Returns true if set_start_time() was called, indicating the client
